+++ wtfix/apps/tests/test_admin.py
@@ -112,7 +112,7 @@ class TestHeartbeatApp:
                     ),
                     0.1,
                 )
-            except asyncio.TimeoutError:
+            except asyncio.exceptions.TimeoutError:
                 pass
 
             assert check.call_count == 0
@@ -133,7 +133,7 @@ class TestHeartbeatApp:
                     ),
                     0.1,
                 )
-            except asyncio.TimeoutError:
+            except asyncio.exceptions.TimeoutError:
                 pass
 
             assert check.call_count > 1
@@ -154,7 +154,7 @@ class TestHeartbeatApp:
                 ),
                 0.1,
             )
-        except asyncio.TimeoutError:
+        except asyncio.exceptions.TimeoutError:
             pass
 
         assert not zero_heartbeat_app._server_not_responding.is_set()